refactored from Vu comments
