+++ packages/anon-aadhaar-pcd/src/index.ts
@@ -2,4 +2,4 @@ export * from './types'
 export * from './pcd'
 export * from './utils'
 export * from './constants'
-export * from './extractWitness'
+export * from './generateArgs'
